chore(server): move agenda toolset module
