+++ packages/server/src/init.ts
@@ -7,9 +7,9 @@ import * as Log from '@/lib/log.js';
 import { refreshMcpToolsets } from '@/mcp/tool-executor.js';
 import { startMeetingDetection } from '@/recordings/meeting-detection.js';
 import { startScheduler } from '@/scheduler/runtime.js';
-import { createAgendaToolset } from '@/tools/core/agenda.js';
 import { createRecordingsToolset } from '@/tools/core/recordings.js';
 import { registerProviderToolsets } from '@/tools/providers/index.js';
+import { createAgendaToolset } from '@/tools/toolsets/agenda.js';
 import { registerToolset } from '@/tools/toolsets/registry.js';
 import { createSessionHistoryToolset } from '@/tools/toolsets/session-history.js';
